Rename rename container to rename instance
