+++ src/classes/Tools/Instances/RenameInstance.php
@@ -1,18 +1,18 @@
 <?php
-namespace dhope0000\LXDClient\Tools\Containers;
+namespace dhope0000\LXDClient\Tools\Instances;
 
 use dhope0000\LXDClient\Model\Client\LxdClient;
 
-class RenameContainer
+class RenameInstance
 {
     public function __construct(LxdClient $lxdClient)
     {
         $this->lxdClient = $lxdClient;
     }
 
-    public function rename(int $hostId, string $container, string $newContainer)
+    public function rename(int $hostId, string $instance, string $newInstance)
     {
         $client = $this->lxdClient->getANewClient($hostId);
-        return $client->instances->rename($container, $newContainer, true);
+        return $client->instances->rename($instance, $newInstance, true);
     }
 }
